+++ cpp/bindings/toppra_int.cpp
@@ -7,7 +7,7 @@
 
 #include <bindings.hpp>
 #include <string>
-#include <toppra/constraint.hpp>
+#include <toppra/algorithm.hpp>
 #include <toppra/geometric_path.hpp>
 #include <toppra/toppra.hpp>
 
@@ -16,26 +16,40 @@ namespace py = pybind11;
 namespace toppra {
 namespace python {
 
+void exposePaths(py::module m);
+void exposeConstraints(py::module m);
+
 PYBIND11_MODULE(toppra_int, m) {
   m.doc() = "toppra C++ bindings (internal)";
-  py::enum_<toppra::DiscretizationType>(m, "DiscretizationType")
-      .value("Collocation", toppra::DiscretizationType::Collocation)
-      .value("Interpolation", toppra::DiscretizationType::Interpolation)
+
+  exposePaths(m);
+  exposeConstraints(m);
+
+  // algorithm
+  py::enum_<toppra::ReturnCode>(m, "ReturnCode")
+      .value("OK", toppra::ReturnCode::OK)
+      .value("ERR_UNKNOWN", toppra::ReturnCode::ERR_UNKNOWN)
+      .value("ERR_UNINITIALIZED", toppra::ReturnCode::ERR_UNINITIALIZED)
+      .value("ERR_FAIL_FEASIBLE", toppra::ReturnCode::ERR_FAIL_FEASIBLE)
+      .value("ERR_FAIL_CONTROLLABLE", toppra::ReturnCode::ERR_FAIL_CONTROLLABLE)
+      .value("ERR_FAIL_FORWARD_PASS", toppra::ReturnCode::ERR_FAIL_FORWARD_PASS)
       .export_values();
 
-  py::class_<PyPiecewisePolyPath>(m, "PiecewisePolyPath")
+  py::class_<toppra::ParametrizationData>(m, "ParametrizationData")
       .def(py::init<>())
-      .def(py::init<const toppra::Matrices&, std::vector<toppra::value_type>>())
-      .def("eval_single", &PyPiecewisePolyPath::eval_single)
-      .def("eval", &PyPiecewisePolyPath::eval)
-      .def("serialize", &PyPiecewisePolyPath::serialize)
-      .def("deserialize", &PyPiecewisePolyPath::deserialize)
-      .def("__call__", &PyPiecewisePolyPath::eval, py::arg("xs"), py::arg("order") = 0)
-      .def("__str__", &PyPiecewisePolyPath::__str__)
-      .def("__repr__", &PyPiecewisePolyPath::__repr__)
-      .def_property_readonly("dof", &PyPiecewisePolyPath::dof)
-      .def_property_readonly("path_interval", &PyPiecewisePolyPath::pathInterval)
-      .def_static("constructHermite", &PyPiecewisePolyPath::constructHermite);
+      .def_readwrite("gridpoints", &toppra::ParametrizationData::gridpoints)
+      .def_readwrite("parametrization", &toppra::ParametrizationData::parametrization)
+      .def_readwrite("controllable_sets",
+                     &toppra::ParametrizationData::controllable_sets)
+      .def_readwrite("feasible_sets", &toppra::ParametrizationData::feasible_sets)
+      .def_readwrite("ret_code", &toppra::ParametrizationData::ret_code);
+
+  py::class_<algorithm::TOPPRA>(m, "TOPPRA")
+      .def(py::init<LinearConstraintPtrs, const GeometricPathPtr &>())
+      .def("computePathParametrization", &algorithm::TOPPRA::computePathParametrization,
+           py::arg("vel_start") = 0, py::arg("vel_end") = 0)
+      .def("setN", &algorithm::TOPPRA::setN)
+      .def_property_readonly("parametrizationData", &algorithm::TOPPRA::getParameterizationData);
 }
 }  // namespace python
 }  // namespace toppra